requirements.txt
requests~=2.32.3
psycopg2-binary~=2.9.10
tqdm~=4.67.1
dotenv~=0.9.9
python-dotenv~=1.1.0
python-dateutil~=2.9.0.post0